Clarified contributing guideline fork.
Many first contributors have been confused about which branch they should pull off of, as DNet Labs' repo development is usually done on the release branch rather than the dev branch that DNet development used to be. As such, a change in the wording of this should probably be justified.

Also added a link to the dnet discord server where the development discussion channel is mentioned. Hyperlinks are cool!